migrate getExtensionInstallationLinks from the extensionCommunication module to the extensionInstallationLinks module
